refactor(server): split projects.test.ts into focused handlers test files (warren-3fcd)
