+++ .mulch/expertise/server.jsonl
@@ -84,3 +84,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:29:35.665Z","evidence":{"commit":"bf1bbc8819f28805972385f605bcb3512c92315b"},"dir_anchors":["src/server/handlers"],"content":"Split src/server/handlers/runs.ts into runs/ subdirectory domain modules (dispatch, lifecycle, events, preview, pause-resume) and runs.test.ts into modular test files (runs.dispatch.test, runs.interactive.test, runs.events.test) with shared test helpers (warren-6566)","id":"mx-cad464"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:37:50.242Z","evidence":{"seeds":"warren-64f6","commit":"709ddbf75aede9ca81ee9ae40d4f60f7253f6fc3"},"content":"Split src/server/handlers/plan-runs.test.ts (1318 lines) into plan-runs.test-helpers.ts and five focused test files under 500 lines: create, plot, get, cancel, events. Removed old monolithic test file and updated file-size budget.","id":"mx-04acb7"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:44:25.942Z","evidence":{"seeds":"warren-d951","commit":"9ff97b9cb04ef33ff4d2f8d7b8cefefa6fde446d"},"content":"Split src/server/handlers/runs.preview.test.ts into three focused files (runs.preview-launch.test.ts, runs.preview-proxy.test.ts, runs.preview-eviction.test.ts) and a shared runs.preview-test-helpers.ts helper, removing the old monolithic file and lowering its budget.","id":"mx-2db6bb"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:52:26.145Z","evidence":{"seeds":"warren-3fcd","commit":"a5b87be6cc271d56ca8e641ce15d6175640b22c2"},"content":"Split projects.test.ts into five focused files under server/handlers to keep per-file line count under the budget","id":"mx-66f7d2"}
