+++ scripts/file-size-budgets.json
@@ -27,7 +27,7 @@
 		"src/server/bridges.test.ts": 647,
 		"src/server/bridges.ts": 655,
 		"src/server/handlers/agents.test.ts": 533,
-		"src/server/handlers/projects.test.ts": 938,
+		"src/server/handlers/projects.settings.test.ts": 685,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
 		"src/server/handlers/plots.list.test.ts": 700,
 		"src/server/handlers/plots.test.ts": 2442,
