+++ src/server/handlers/projects.refresh.test.ts
@@ -1,0 +1,120 @@
+import { afterEach, beforeEach, describe, expect, test } from "bun:test";
+import { BurrowClient } from "../../burrow-client/index.ts";
+import { openDatabase, type WarrenDb } from "../../db/client.ts";
+import { createRepos, type Repos } from "../../db/repos/index.ts";
+import { NO_AUTH } from "../auth.ts";
+import { startServer } from "../server.ts";
+import type { ServeHandle, ServerDeps } from "../types.ts";
+import { depsFor, silentLogger, stub, tcpUrl } from "./projects.test-helpers.ts";
+
+describe("POST /projects/:id/refresh — git fetch + hard reset", () => {
+	let db: WarrenDb;
+	let repos: Repos;
+	let handle: ServeHandle | null = null;
+	let projectLocalPath = "";
+	let projectId = "";
+
+	beforeEach(async () => {
+		db = await openDatabase({ path: ":memory:" });
+		repos = createRepos(db);
+
+		const { mkdtemp } = await import("node:fs/promises");
+		const { tmpdir } = await import("node:os");
+		const { join } = await import("node:path");
+		projectLocalPath = await mkdtemp(join(tmpdir(), "warren-refresh-proj-"));
+
+		const row = await repos.projects.create({
+			gitUrl: "https://github.com/x/y.git",
+			localPath: projectLocalPath,
+			defaultBranch: "main",
+		});
+		projectId = row.id;
+	});
+
+	afterEach(async () => {
+		if (handle) {
+			await handle.stop();
+			handle = null;
+		}
+		await db.close();
+	});
+
+	test("refreshes the clone, stamps lastFetchedAt + lastHeadSha, returns 200", async () => {
+		const burrowClient = new BurrowClient({
+			config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
+			fetch: stub(async () => new Response("{}", { status: 200 })),
+		});
+		const deps = await depsFor(repos, burrowClient);
+		handle = startServer(deps, {
+			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
+			auth: NO_AUTH,
+			logger: silentLogger,
+		});
+
+		const res = await fetch(`${tcpUrl(handle)}/projects/${projectId}/refresh`, {
+			method: "POST",
+		});
+		expect(res.status).toBe(200);
+		const body = (await res.json()) as {
+			project: { id: string; lastHeadSha: string | null; lastFetchedAt: string | null };
+			headSha: string;
+			ref: string;
+		};
+		expect(body.project.id).toBe(projectId);
+		expect(body.headSha).toBe("deadbeef".repeat(5));
+		expect(body.ref).toBe("main");
+		expect(body.project.lastHeadSha).toBe("deadbeef".repeat(5));
+		expect(body.project.lastFetchedAt).not.toBeNull();
+	});
+
+	test("forwards an explicit ref into the refresh", async () => {
+		const burrowClient = new BurrowClient({
+			config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
+			fetch: stub(async () => new Response("{}", { status: 200 })),
+		});
+		const seenRefs: string[] = [];
+		const deps: ServerDeps = {
+			...(await depsFor(repos, burrowClient)),
+			spawn: async (cmd) => {
+				if (cmd[1] === "checkout") {
+					seenRefs.push(cmd[3] ?? "");
+				}
+				if (cmd[1] === "rev-parse") {
+					return { stdout: "abc1234".padEnd(40, "0"), stderr: "", exitCode: 0 };
+				}
+				return { stdout: "", stderr: "", exitCode: 0 };
+			},
+		};
+		handle = startServer(deps, {
+			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
+			auth: NO_AUTH,
+			logger: silentLogger,
+		});
+
+		const res = await fetch(`${tcpUrl(handle)}/projects/${projectId}/refresh`, {
+			method: "POST",
+			headers: { "content-type": "application/json" },
+			body: JSON.stringify({ ref: "feature/x" }),
+		});
+		expect(res.status).toBe(200);
+		expect(seenRefs).toEqual(["feature/x"]);
+	});
+
+	test("returns 404 for an unknown project id", async () => {
+		const burrowClient = new BurrowClient({
+			config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
+			fetch: stub(async () => new Response("{}", { status: 200 })),
+		});
+		const deps = await depsFor(repos, burrowClient);
+		handle = startServer(deps, {
+			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
+			auth: NO_AUTH,
+			logger: silentLogger,
+		});
+
+		const res = await fetch(`${tcpUrl(handle)}/projects/prj_doesnotexist/refresh`, {
+			method: "POST",
+		});
+		expect(res.status).toBe(404);
+	});
+});
