+++ src/server/handlers/projects.settings.test.ts
@@ -1,270 +1,11 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { BurrowClient, BurrowClientPool } from "../../burrow-client/index.ts";
+import { BurrowClient } from "../../burrow-client/index.ts";
 import { openDatabase, type WarrenDb } from "../../db/client.ts";
 import { createRepos, type Repos } from "../../db/repos/index.ts";
-import { RunEventBroker } from "../../runs/index.ts";
 import { NO_AUTH } from "../auth.ts";
-import { createBridgeRegistry } from "../bridges.ts";
 import { startServer } from "../server.ts";
 import type { BridgeRegistry, ServeHandle, ServerDeps } from "../types.ts";
-
-/**
- * Build a single-worker `BurrowClientPool` from a stubbed `BurrowClient`
- * so `POST /runs` and `POST /projects/:id/triggers/:triggerId/run` can
- * route through `spawnRun`'s placement seam (warren-39c3). Upserts the
- * synthetic `local` worker row so `placeForProject` has a healthy
- * candidate.
- */
-async function poolFor(repos: Repos, client: BurrowClient): Promise<BurrowClientPool> {
-	await repos.workers.upsert({ name: "local", url: "unix:///tmp/x.sock" });
-	const pool = new BurrowClientPool({ repos });
-	pool.register("local", client);
-	return pool;
-}
-
-const silentLogger = {
-	info() {},
-	warn() {},
-	error() {},
-};
-
-function stub(
-	impl: (input: URL | RequestInfo, init?: RequestInit) => Promise<Response>,
-): typeof fetch {
-	return impl as unknown as typeof fetch;
-}
-
-interface BurrowFixture {
-	burrowId: string;
-	burrowRunId: string;
-	workspacePath: string;
-}
-
-function makeBurrowClient(
-	fix: BurrowFixture,
-	calls: { method: string; path: string; body: unknown }[],
-): BurrowClient {
-	return new BurrowClient({
-		config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
-		fetch: stub(async (input, init) => {
-			const url = new URL(String(input), "http://localhost");
-			const path = url.pathname;
-			const method = init?.method ?? "GET";
-			const reqBody = typeof init?.body === "string" ? JSON.parse(init.body) : undefined;
-			calls.push({ method, path, body: reqBody });
-			if (method === "POST" && path === "/burrows") {
-				const burrow = {
-					id: fix.burrowId,
-					name: "burrow",
-					kind: "task",
-					projectRoot: "/data/projects/x/y",
-					branch: "main",
-					baseBranch: "main",
-					originUrl: "https://github.com/x/y.git",
-					workspacePath: fix.workspacePath,
-					provider: "local",
-					sandbox: { network: "open" },
-					state: "running",
-					createdAt: "2026-05-08T12:00:00Z",
-					updatedAt: "2026-05-08T12:00:00Z",
-				};
-				return new Response(JSON.stringify(burrow), {
-					status: 201,
-					headers: { "content-type": "application/json" },
-				});
-			}
-			if (method === "POST" && path === `/burrows/${fix.burrowId}/runs`) {
-				const run = {
-					id: fix.burrowRunId,
-					burrowId: fix.burrowId,
-					agentId: "refactor-bot",
-					prompt: "hello",
-					resumeOfRunId: null,
-					state: "queued",
-					exitCode: null,
-					errorMessage: null,
-					metadataJson: null,
-					queuedAt: "2026-05-08T12:00:01Z",
-					startedAt: null,
-					completedAt: null,
-				};
-				return new Response(JSON.stringify(run), {
-					status: 201,
-					headers: { "content-type": "application/json" },
-				});
-			}
-			return new Response(
-				JSON.stringify({ error: { code: "not_found", message: `unmatched ${method} ${path}` } }),
-				{
-					status: 404,
-					headers: { "content-type": "application/json" },
-				},
-			);
-		}),
-	});
-}
-
-async function depsFor(
-	repos: Repos,
-	burrowClient: BurrowClient,
-	bridges?: BridgeRegistry,
-	extras?: { plotResolver?: import("../../plots/index.ts").PlotResolver },
-): Promise<ServerDeps> {
-	const broker = new RunEventBroker();
-	const burrowClientPool = await poolFor(repos, burrowClient);
-	return {
-		repos,
-		burrowClientPool,
-		broker,
-		bridges:
-			bridges ??
-			createBridgeRegistry({
-				repos,
-				broker,
-				burrowClientPool,
-				bridge: async () => ({ written: 0, skipped: 0, errored: false }),
-			}),
-		canopyConfig: {
-			repoUrl: "https://example/agents.git",
-			localDir: "/tmp/cn",
-			cnBinary: "cn",
-			gitBinary: "git",
-		},
-		projectsConfig: { root: "/tmp/projects", gitBinary: "git" },
-		logger: silentLogger,
-		uiDistDir: null,
-		// No-op spawn so the project-refresh path inside `POST /runs` and
-		// `POST /projects/:id/refresh` doesn't shell out to real `git`
-		// against tmpdir paths the test never populated. Tests that need
-		// to assert on spawn calls override `deps.spawn` directly.
-		spawn: async (cmd) => {
-			if (cmd[1] === "rev-parse") {
-				return { stdout: "deadbeef".repeat(5), stderr: "", exitCode: 0 };
-			}
-			return { stdout: "", stderr: "", exitCode: 0 };
-		},
-		...(extras?.plotResolver !== undefined ? { plotResolver: extras.plotResolver } : {}),
-	};
-}
-
-function tcpUrl(handle: ServeHandle): string {
-	if (handle.transport.kind !== "tcp") throw new Error("expected tcp transport");
-	return `http://${handle.transport.hostname}:${handle.transport.port}`;
-}
-
-/* Projects handlers (extracted from handlers.test.ts, warren-599c / pl-9088 step 3). */
-
-describe("POST /projects/:id/refresh — git fetch + hard reset", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-	let projectLocalPath = "";
-	let projectId = "";
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-
-		const { mkdtemp } = await import("node:fs/promises");
-		const { tmpdir } = await import("node:os");
-		const { join } = await import("node:path");
-		projectLocalPath = await mkdtemp(join(tmpdir(), "warren-refresh-proj-"));
-
-		const row = await repos.projects.create({
-			gitUrl: "https://github.com/x/y.git",
-			localPath: projectLocalPath,
-			defaultBranch: "main",
-		});
-		projectId = row.id;
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	test("refreshes the clone, stamps lastFetchedAt + lastHeadSha, returns 200", async () => {
-		const burrowClient = new BurrowClient({
-			config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
-			fetch: stub(async () => new Response("{}", { status: 200 })),
-		});
-		const deps = await depsFor(repos, burrowClient);
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/projects/${projectId}/refresh`, {
-			method: "POST",
-		});
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as {
-			project: { id: string; lastHeadSha: string | null; lastFetchedAt: string | null };
-			headSha: string;
-			ref: string;
-		};
-		expect(body.project.id).toBe(projectId);
-		expect(body.headSha).toBe("deadbeef".repeat(5));
-		expect(body.ref).toBe("main");
-		expect(body.project.lastHeadSha).toBe("deadbeef".repeat(5));
-		expect(body.project.lastFetchedAt).not.toBeNull();
-	});
-
-	test("forwards an explicit ref into the refresh", async () => {
-		const burrowClient = new BurrowClient({
-			config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
-			fetch: stub(async () => new Response("{}", { status: 200 })),
-		});
-		const seenRefs: string[] = [];
-		const deps: ServerDeps = {
-			...(await depsFor(repos, burrowClient)),
-			spawn: async (cmd) => {
-				if (cmd[1] === "checkout") {
-					seenRefs.push(cmd[3] ?? "");
-				}
-				if (cmd[1] === "rev-parse") {
-					return { stdout: "abc1234".padEnd(40, "0"), stderr: "", exitCode: 0 };
-				}
-				return { stdout: "", stderr: "", exitCode: 0 };
-			},
-		};
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/projects/${projectId}/refresh`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({ ref: "feature/x" }),
-		});
-		expect(res.status).toBe(200);
-		expect(seenRefs).toEqual(["feature/x"]);
-	});
-
-	test("returns 404 for an unknown project id", async () => {
-		const burrowClient = new BurrowClient({
-			config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
-			fetch: stub(async () => new Response("{}", { status: 200 })),
-		});
-		const deps = await depsFor(repos, burrowClient);
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/projects/prj_doesnotexist/refresh`, {
-			method: "POST",
-		});
-		expect(res.status).toBe(404);
-	});
-});
+import { depsFor, makeBurrowClient, silentLogger, stub, tcpUrl } from "./projects.test-helpers.ts";
 
 describe("GET /projects/:id/warren-config — per-project .warren/ envelope (warren-435b)", () => {
 	let db: WarrenDb;
