+++ src/server/handlers/projects.test-helpers.ts
@@ -1,0 +1,149 @@
+import { BurrowClient, BurrowClientPool } from "../../burrow-client/index.ts";
+import type { Repos } from "../../db/repos/index.ts";
+import { RunEventBroker } from "../../runs/index.ts";
+import { createBridgeRegistry } from "../bridges.ts";
+import type { BridgeRegistry, ServeHandle, ServerDeps } from "../types.ts";
+
+/**
+ * Build a single-worker `BurrowClientPool` from a stubbed `BurrowClient`
+ * so `POST /runs` and `POST /projects/:id/triggers/:triggerId/run` can
+ * route through `spawnRun`'s placement seam (warren-39c3). Upserts the
+ * synthetic `local` worker row so `placeForProject` has a healthy
+ * candidate.
+ */
+export async function poolFor(repos: Repos, client: BurrowClient): Promise<BurrowClientPool> {
+	await repos.workers.upsert({ name: "local", url: "unix:///tmp/x.sock" });
+	const pool = new BurrowClientPool({ repos });
+	pool.register("local", client);
+	return pool;
+}
+
+export const silentLogger = {
+	info() {},
+	warn() {},
+	error() {},
+};
+
+export function stub(
+	impl: (input: URL | RequestInfo, init?: RequestInit) => Promise<Response>,
+): typeof fetch {
+	return impl as unknown as typeof fetch;
+}
+
+export interface BurrowFixture {
+	burrowId: string;
+	burrowRunId: string;
+	workspacePath: string;
+}
+
+export function makeBurrowClient(
+	fix: BurrowFixture,
+	calls: { method: string; path: string; body: unknown }[],
+): BurrowClient {
+	return new BurrowClient({
+		config: { transport: { kind: "unix", path: "/tmp/x.sock" } },
+		fetch: stub(async (input, init) => {
+			const url = new URL(String(input), "http://localhost");
+			const path = url.pathname;
+			const method = init?.method ?? "GET";
+			const reqBody = typeof init?.body === "string" ? JSON.parse(init.body) : undefined;
+			calls.push({ method, path, body: reqBody });
+			if (method === "POST" && path === "/burrows") {
+				const burrow = {
+					id: fix.burrowId,
+					name: "burrow",
+					kind: "task",
+					projectRoot: "/data/projects/x/y",
+					branch: "main",
+					baseBranch: "main",
+					originUrl: "https://github.com/x/y.git",
+					workspacePath: fix.workspacePath,
+					provider: "local",
+					sandbox: { network: "open" },
+					state: "running",
+					createdAt: "2026-05-08T12:00:00Z",
+					updatedAt: "2026-05-08T12:00:00Z",
+				};
+				return new Response(JSON.stringify(burrow), {
+					status: 201,
+					headers: { "content-type": "application/json" },
+				});
+			}
+			if (method === "POST" && path === `/burrows/${fix.burrowId}/runs`) {
+				const run = {
+					id: fix.burrowRunId,
+					burrowId: fix.burrowId,
+					agentId: "refactor-bot",
+					prompt: "hello",
+					resumeOfRunId: null,
+					state: "queued",
+					exitCode: null,
+					errorMessage: null,
+					metadataJson: null,
+					queuedAt: "2026-05-08T12:00:01Z",
+					startedAt: null,
+					completedAt: null,
+				};
+				return new Response(JSON.stringify(run), {
+					status: 201,
+					headers: { "content-type": "application/json" },
+				});
+			}
+			return new Response(
+				JSON.stringify({ error: { code: "not_found", message: `unmatched ${method} ${path}` } }),
+				{
+					status: 404,
+					headers: { "content-type": "application/json" },
+				},
+			);
+		}),
+	});
+}
+
+export async function depsFor(
+	repos: Repos,
+	burrowClient: BurrowClient,
+	bridges?: BridgeRegistry,
+	extras?: { plotResolver?: import("../../plots/index.ts").PlotResolver },
+): Promise<ServerDeps> {
+	const broker = new RunEventBroker();
+	const burrowClientPool = await poolFor(repos, burrowClient);
+	return {
+		repos,
+		burrowClientPool,
+		broker,
+		bridges:
+			bridges ??
+			createBridgeRegistry({
+				repos,
+				broker,
+				burrowClientPool,
+				bridge: async () => ({ written: 0, skipped: 0, errored: false }),
+			}),
+		canopyConfig: {
+			repoUrl: "https://example/agents.git",
+			localDir: "/tmp/cn",
+			cnBinary: "cn",
+			gitBinary: "git",
+		},
+		projectsConfig: { root: "/tmp/projects", gitBinary: "git" },
+		logger: silentLogger,
+		uiDistDir: null,
+		// No-op spawn so the project-refresh path inside `POST /runs` and
+		// `POST /projects/:id/refresh` doesn't shell out to real `git`
+		// against tmpdir paths the test never populated. Tests that need
+		// to assert on spawn calls override `deps.spawn` directly.
+		spawn: async (cmd) => {
+			if (cmd[1] === "rev-parse") {
+				return { stdout: "deadbeef".repeat(5), stderr: "", exitCode: 0 };
+			}
+			return { stdout: "", stderr: "", exitCode: 0 };
+		},
+		...(extras?.plotResolver !== undefined ? { plotResolver: extras.plotResolver } : {}),
+	};
+}
+
+export function tcpUrl(handle: ServeHandle): string {
+	if (handle.transport.kind !== "tcp") throw new Error("expected tcp transport");
+	return `http://${handle.transport.hostname}:${handle.transport.port}`;
+}
